move tests to irm submodule
