+++ doubleml/irm/tests/_utils_cvar_manual.py
@@ -2,8 +2,8 @@ import numpy as np
 from sklearn.base import clone
 from sklearn.model_selection import train_test_split, StratifiedKFold
 
-from ._utils import fit_predict_proba, tune_grid_search
-from ..utils._estimation import _dml_cv_predict, _normalize_ipw, _get_bracket_guess, _solve_ipw_score
+from ...tests._utils import fit_predict_proba, tune_grid_search
+from ...utils._estimation import _dml_cv_predict, _normalize_ipw, _get_bracket_guess, _solve_ipw_score
 
 
 def fit_cvar(y, x, d, quantile,
