+++ doubleml/irm/tests/_utils_irm_manual.py
@@ -1,11 +1,11 @@
 import numpy as np
 from sklearn.base import clone, is_classifier
 
-from ._utils_boot import boot_manual, draw_weights
-from ._utils import fit_predict, fit_predict_proba, tune_grid_search
+from ...tests._utils_boot import boot_manual, draw_weights
+from ...tests._utils import fit_predict, fit_predict_proba, tune_grid_search
 
-from ..utils._estimation import _normalize_ipw
-from ..utils._checks import _check_is_propensity
+from ...utils._estimation import _normalize_ipw
+from ...utils._checks import _check_is_propensity
 
 
 def fit_irm(y, x, d,
