+++ doubleml/irm/tests/_utils_qte_manual.py
@@ -3,11 +3,11 @@ from sklearn.base import clone
 import pandas as pd
 from scipy.stats import norm
 
-from ..irm.pq import DoubleMLPQ
-from ..double_ml_data import DoubleMLData
+from ..pq import DoubleMLPQ
+from ...double_ml_data import DoubleMLData
 
-from ._utils_boot import draw_weights
-from ..utils._estimation import _default_kde
+from ...tests._utils_boot import draw_weights
+from ...utils._estimation import _default_kde
 
 
 def fit_qte(y, x, d, quantiles, learner_g, learner_m, all_smpls, n_rep=1, dml_procedure='dml2',
