+++ doubleml/irm/tests/test_irm_no_cross_fit.py
@@ -8,7 +8,7 @@ from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor
 
 import doubleml as dml
 
-from ._utils import draw_smpls
+from ...tests._utils import draw_smpls
 from ._utils_irm_manual import fit_irm, boot_irm
