+++ doubleml/irm/tests/test_irm_with_missings.py
@@ -10,7 +10,7 @@ from xgboost import XGBClassifier, XGBRegressor
 
 import doubleml as dml
 
-from ._utils import draw_smpls
+from ...tests._utils import draw_smpls
 from ._utils_irm_manual import fit_irm, boot_irm
