+++ doubleml/irm/tests/test_lpq_external_predictions.py
@@ -5,7 +5,7 @@ from sklearn.linear_model import LogisticRegression
 from doubleml import DoubleMLLPQ, DoubleMLData
 from doubleml.datasets import make_iivm_data
 from doubleml.utils import DMLDummyClassifier
-from ._utils import draw_smpls
+from ...tests._utils import draw_smpls
 
 
 @pytest.fixture(scope="module", params=["dml1", "dml2"])
